Add system blueprint with health and stats endpoints
